requirements.txt
pandas
psycopg2-binary
sqlalchemy
jupyter
matplotlib
seaborn
python-dotenv